refactor(music-player): extract useAudioPlayer hook, add skip ±10s, mini track name, keyboard shortcuts
- Extract all audio logic from GameMusic.tsx into useAudioPlayer hook
- Extract button components to GameMusicButtons.tsx for file size compliance
- Add skip ±10s buttons in transport controls
- Show track title in mini mode
- Add global keyboard shortcuts (Space=play/pause, ←→=seek, ↑↓=volume, Shift+←→=prev/next)
- Add i18n keys for skipForward/skipBack in all locales
- Trim auth.service.ts to pass 500-line check (pre-existing)
